% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_default_rate.R
\name{fit_default_rate}
\alias{fit_default_rate}
\title{This is default rate model function}
\usage{
fit_default_rate(default_rates)
}
\arguments{
\item{default_rates}{training data}
}
\description{
This function allows you to create a model to predict the 
default rate for student loan application.
}
\examples{
fit_default_rate()
}
\keyword{default_rate}
