% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{This is load data function}
\usage{
load_data()
}
\description{
This function allows you to load the training data
required for student loan default prediction model.
}
\examples{
load_data()
}
\keyword{default_rate}
