% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{This is load packages function}
\usage{
load_packages()
}
\description{
This function allows you to load all the required packages
required for student loan default prediction model.
}
\examples{
load_packages()
}
\keyword{default_rate}
