% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_default_rate.R
\name{predict_default_rate}
\alias{predict_default_rate}
\title{This is default rate prediction function}
\usage{
predict_default_rate(test_data)
}
\arguments{
\item{test_data}{student loan application parameters}
}
\description{
This function allows you to predict the 
default rate for student loan application.
}
\examples{
predict_default_rate()
}
\keyword{default_rate}
