% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBackground.R
\name{setBackground}
\alias{setBackground}
\title{Set value background cue.}
\usage{
setBackground(value)
}
\arguments{
\item{value}{A string specifying the background cue. 
NULL or FALSE or a number < 0 indicates to remove the background cue, 
whereas the values TRUE or a number > 0 set the value to "***", and a 
string can specify the value for the background cue.}
}
\value{
No return value
}
\description{
Set value background cue.
}
